requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.51.2
supabase==2.14.0
python-dateutil==2.9.0
tqdm==4.67.1
python-dotenv==1.0.1 